More work on value objects

+++ BusinessAssociates.Domain/Events/Events.cs
@@ -17,37 +17,37 @@ namespace BusinessAssociates.Domain
 
         public class InternalAssociateDUNSNumberUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public int DUNSNumber { get; set; }
         }
 
         public class InternalAssociateTypeUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public int InternalAssociateType { get; set; }
         }
 
         public class InternalAssociateLongNameUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string LongName { get; set; }
         }
 
         public class InternalAssociateIsParentUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public bool IsParent { get; set; }
         }
 
         public class InternalAssociateStatusUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public int Status { get; set; }
         }
 
         public class InternalAssociateShortNameUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string ShortName { get; set; }
         }
 
@@ -71,30 +71,31 @@ namespace BusinessAssociates.Domain
 
         public class ExternalAssociateTypeUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public int ExternalAssociateType { get; set; }
         }
 
         public class ExternalAssociateLongNameUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string LongName { get; set; }
         }
 
         public class ExternalAssociateIsParentUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public bool IsParent { get; set; }
         }
+
         public class ExternalAssociateStatusUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public int Status { get; set; }
         }
 
         public class ExternalAssociateShortNameUpdated
         {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string ShortName { get; set; }
         }
     }
